+++ tsconfig.json
@@ -17,7 +17,7 @@
         "baseUrl": ".",
         "paths": {
             "@/lib/*": ["lib/*"],
-            "@/utils/*": ["utils/*"],
+            "@/helpers/*": ["helpers/*"],
             "@/types/*": ["types/*"],
             "@/componenets/*": ["componenets/*"]
         }
